requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
python-dotenv==1.0.1
SQLAlchemy==2.0.28
Werkzeug==3.0.1
mysqlclient==2.2.4
cloud-sql-python-connector==1.7.0
PyMySQL==1.1.0 